cleanup

+++ core/src/main/java/tech/tablesaw/table/TableSlice.java
@@ -162,7 +162,7 @@ public class TableSlice extends Relation implements IntIterable {
      */
     public double reduce(String numberColumnName, NumericAggregateFunction function) {
         NumberColumn column = table.numberColumn(numberColumnName);
-        return (Double) function.summarize(column.where(selection));
+        return function.summarize(column.where(selection));
     }
 
     /**
